LayoutTests/media/modern-media-controls/layout-item/layout-item-expected.txt

Testing the LayoutItem class.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


Creating a LayoutItem and setting a layout delegate after construction.
PASS itemWithNoParameter.layoutTraits threw exception No layout traits specified via layoutDelegate or overriden layouTraits() getter.
PASS itemWithNoParameter.layoutTraits is an instance of MacOSLayoutTraits
PASS itemWithNoParameter.layoutTraits.isFullscreen is true

Creating a LayoutItem with a layout delegate set at construction.
PASS itemWithNoParameter.layoutTraits is an instance of MacOSLayoutTraits
PASS itemWithNoParameter.layoutTraits.isFullscreen is true
PASS successfullyParsed is true

TEST COMPLETE
